:recycle: move submission delete tests mixin in common module
