+++ django-data/image/submissions/urls.py
@@ -30,18 +30,23 @@ urlpatterns = [
     url(r'^(?P<pk>[-\w]+)/reload/$',
         views.ReloadSubmissionView.as_view(),
         name='reload'),
+
     url(r'^(?P<pk>[-\w]+)/delete/$',
         views.DeleteSubmissionView.as_view(),
         name='delete'),
+
     url(r'^(?P<pk>[-\w]+)/delete_animals/$',
         views.DeleteAnimalsView.as_view(),
         name='delete_animals'),
+
     url(r'^(?P<pk>[-\w]+)/delete_samples/$',
         views.DeleteSamplesView.as_view(),
         name='delete_samples'),
+
     url(r'^(?P<pk>[-\w]+)/validation_summary/(?P<type>[\w]+)/$',
         views.SubmissionValidationSummaryView.as_view(),
         name='validation_summary'),
+
     url(r'^(?P<pk>[-\w]+)/batch_delete/(?P<type>[\w]+)/$',
         views.BatchDelete.as_view(),
         name='batch_delete')
