+++ django-data/image/validation/models.py
@@ -81,6 +81,7 @@ class ValidationSummary(models.Model):
 
     def reset(self):
         """Sets all counts to 0, except all_count field"""
+
         self.pass_count = 0
         self.warning_count = 0
         self.error_count = 0
@@ -88,3 +89,5 @@ class ValidationSummary(models.Model):
         self.validation_known_count = 0
         self.messages = list()
         self.save()
+
+        # HINT: why not reset all count here ?
